+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Added support for the `Pager.refresh` method to accept an optional `refreshKey` parameter.
+
 ## 0.1.0
 
 - Initial Release.